refactor(server): consolidate test fixtures into conftest hierarchy
make_adapter was duplicated between test_processors.py and cf/conftest.py;
moved to tests/unit/conftest.py as the single source. make_ioc and a new
make_channel (generalising make_active_channel with name, recceiver_id,
and active params) now live in tests/unit/cf/conftest.py for shared use
across all cf tests.
