+++ server/tests/unit/conftest.py
@@ -1,0 +1,14 @@
+from configparser import ConfigParser
+
+from recceiver.processors import ConfigAdapter
+
+
+def make_adapter(section: str = "cf", values: dict = None, env: dict = None) -> ConfigAdapter:
+    parser = ConfigParser()
+    parser.add_section(section)
+    for key, value in (values or {}).items():
+        parser.set(section, key, str(value))
+    adapter = ConfigAdapter(parser, section)
+    if env:
+        adapter.env_vars = env
+    return adapter
